Fix a bunch of static analysis warnings, add more tests for zip, and fix some bugs in zip [ci skip]

+++ include/parlay/hash_table.h
@@ -64,7 +64,7 @@ class hashtable {
   // Size is the maximum number of values the hash table will hold.
   // Overfilling the table could put it into an infinite loop.
   hashtable(size_t size, HASH hashF, double load = 1.5)
-    : m(100 + static_cast<size_t>(load * size)),
+    : m(100 + static_cast<size_t>(load * static_cast<double>(size))),
       empty(hashF.empty()),
       hashStruct(hashF),
       TA(sequence<eType>(m, empty)) {
